hermes-trading-post/requirements_gpu.txt
# ModernGL GPU Chart Dependencies
moderngl>=5.8.0
pygame>=2.0.0
Pillow>=9.0.0
numpy>=1.20.0

# Existing dependencies (keep these)
dash>=2.17.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0
pandas>=1.5.0
requests>=2.28.0
websocket-client>=1.6.0
